Extracted endpoint base URL
